Show the new shared list
1. when refresh button pressed, the new list show up
2. selecting among the shared book list, then move to mainpage with that account book.

+++ js/analysis.js
@@ -40,7 +40,7 @@ function Analysis(goldlucksdb){
         amount = array[i].amount;
         $('#analyList').append('<li>'+cat+'<p class="ui-li-aside">'+amount+'$</p>');
         $('#analyList').listview('refresh');
-        sum += parseInt(amount);
+        sum += parseFloat(amount);
       }
       $('#analyTotal').html(sum+'$');
     },
